refactor(settings): give the settings file its own package
The settings file sat under cmdutil because that is where it happened to
land when it left the command tree. Nothing about it is a command
utility: it is the file a repository commits its linting policy in, with
a shape, a loader, a JSON Schema and a rule selection of its own, and it
is read by the binary and describable without a command anywhere in
sight.

pkg/settings, then, next to the other things the linter is made of --
pkg/schema, pkg/ruleset, pkg/lint. cmdutil is what is left when nothing
that stands on its own is in it: the options every command takes, and the
three exit codes.
