+++ pkg/cmd/otelcol-config-lint/configschema/configschema.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/minuk-dev/otelcol-config-lint/pkg/cmdutil/settings"
+	"github.com/minuk-dev/otelcol-config-lint/pkg/settings"
 )
 
 // NewCommand builds "config-schema", which prints the JSON Schema for the
